service-account v0.6.0: Adds helm pre-install hooks to prevent issue (#110)
There is a timing issue/race condition that happens if the deployment/pod that uses the service account is created before the
scc is applied to the service account. The condition will eventually resolve itself but the immediate deployment appears to fail because the scc has not been applied.

The pre-install hooks should change the order in which the resources are created.